pyproject.toml
[tool.poetry]
name = "devops-project"
version = "0.1.0"
description = ""
authors = ["tuan220199 <phamtuan.220199@gmail.com>"]
license = "MIT"
package-mode = false  # Disable package mode to avoid installation issues

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
django-filter = "^24.2"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
